[d3d9] Fix GCC warnings for unsigned->signed comparisons

+++ src/d3d9/d3d9_util.h
@@ -251,10 +251,10 @@ namespace dxvk {
     // Only bother checking x, y as we don't have 3D blits.
     return offsets[1].x < offsets[0].x ||
            offsets[1].y < offsets[0].y ||
-           offsets[0].x < 0             ||
-           offsets[0].y < 0             ||
-           offsets[1].x > extent.width  ||
-           offsets[1].y > extent.height;
+           offsets[0].x < 0 ||
+           offsets[0].y < 0 ||
+           uint32_t(offsets[1].x) > extent.width ||
+           uint32_t(offsets[1].y) > extent.height;
   }
 
   enum D3D9TextureStageStateTypes : uint32_t
